+++ src/todos.ts
@@ -19,9 +19,11 @@ export function getCommentsOnCode(filePath:string) {
     let lineNumber = 0;
 		
     while ((match = singleLineCommentRegex.exec(code)) !== null) {
-				const firstTodoIndex = match[0].trim().toLocaleLowerCase().indexOf('todo');
-				const commentText = match[0].substring(firstTodoIndex + 4).trim(); // Agrega 4 para omitir "TODO"
-				
+
+				const todoWord = /todo\s*:*/gi;
+				const splittedText = match[0].trim().toLocaleLowerCase().split(todoWord);
+				const commentText = splittedText[1];
+
         const lines = code.substring(0, match.index).split('\n');
         lineNumber = lines.length;
 
@@ -30,8 +32,9 @@ export function getCommentsOnCode(filePath:string) {
 
     //? Extract block comments
     while ((match = blockCommentRegex.exec(code)) !== null) {
-				const firstTodoIndex = match[0].trim().toLocaleLowerCase().indexOf('todo');
-				const commentText = match[0].substring(firstTodoIndex + 4).trim(); 
+				const todoWord = /todo\s*:*/gi;
+				const splittedText = match[0].trim().toLocaleLowerCase().split(todoWord);
+				const commentText = splittedText[1];
 				const withoutEndOfBlock = commentText.replace(/([\s\S]-+->|\/\*\s*$|[\s\S]\*+\/|[\s\S]'''|[\s\S]""")/gi, '');
         const lines = code.substring(0, match.index).split('\n');
         lineNumber = lines.length + commentText.split('\n').length - 1;
@@ -62,7 +65,7 @@ export function findTodosInProject(): { file: string, line: number, text: string
     const projectDir = vscode.workspace.rootPath;
 
     if (!projectDir) {
-        vscode.window.showWarningMessage('No se encontró un proyecto abierto.');
+        vscode.window.showWarningMessage('No opened project found.');
         return [];
     }